third_party/blink/web_tests/external/wpt/css/css-tables/html5-table-formatting-1-expected.txt

This is a testharness.js-based test.
[FAIL] Empty tables do not take table-columns into account
  assert_equals: expected 50 but got 200
[FAIL] Empty tables do not take table-rows into account
  assert_equals: expected 50 but got 100
Harness: the test ran to completion.
